fix: TileRepository의 ID 타입을 Tile.id(Long)와 일치시킴
JpaRepository<Tile, Integer>였는데 Tile.id는 Long이라 타입 불일치.
현재는 findById류를 안 써서 무증상이지만 잠복 버그였음.

+++ src/main/java/com/seohamin/pomoland/domain/map/tile/repository/TileRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface TileRepository extends JpaRepository<Tile, Integer> {
+public interface TileRepository extends JpaRepository<Tile, Long> {
     Optional<Tile> findByOwnerIdAndIsSpawnPoint(final Long ownerId, final Boolean isSpawnPoint);
     boolean existsByOwnerIdAndIsSpawnPoint(final Long ownerId, final Boolean isSpawnPoint);
     Optional<Tile> findByXAndY(final Integer x, final Integer y);
